macbook: cumulative changes for last years

+++ algorithm/subarray_min_count_sum.h
@@ -3,30 +3,29 @@
 
 #include "../main/header.h"
 
-// 
+//
 // https://www.hackerrank.com/challenges/subset-sum/problem
 
-// 
+//
 // dynamic programming
-int
-subarray_min_count_sum(const vector<int>& arr, int sum)
+int subarray_min_count_sum(const vector<int>& arr, int sum)
 {
-    auto len = arr.size();
+    auto                len = arr.size();
     vector<vector<int>> dp;
-    dp.resize(len, vector<int>(sum+1, 0));
-    for (int j = 0; j < sum+1; ++j) {
+    dp.resize(len, vector<int>(sum + 1, 0));
+    for (int j = 0; j < sum + 1; ++j) {
         dp[0][j] = (arr[0] >= j ? 1 : -1);
     }
-    for (int i = 1; i < len; ++i) {
-        for (int j = 0; j < sum+1; ++j) {
+    for (size_t i = 1; i < len; ++i) {
+        for (int j = 0; j < sum + 1; ++j) {
             if (arr[i] >= j) {
                 dp[i][j] = 1;
             } else {
                 int v = 0;
                 // doesn't include arr[i] into sum
-                auto v1 = dp[i-1][j];   // -1, or > 0
+                auto v1 = dp[i - 1][j];    // -1, or > 0
                 // include arr[i] into sum
-                auto v2 = dp[i-1][j-arr[i]] + 1;    // 0, or > 0
+                auto v2 = dp[i - 1][j - arr[i]] + 1;    // 0, or > 0
                 if (v1 > 0) {
                     if (v2 > 0) {
                         v = min(v1, v2);
@@ -44,7 +43,7 @@ subarray_min_count_sum(const vector<int>& arr, int sum)
             }
         }
     }
-    return dp[len-1][sum];
+    return dp[len - 1][sum];
 }
 
 #endif
